Fix missing autograd box schema and move schema creation code to components/type_util.py
